+++ demo/nuxt/components/DemoView.vue
@@ -4,7 +4,8 @@ defineProps<{ title: string; body: string }>();
 
 <template>
   <div class="view">
-    <h2>{{ title }}</h2>
-    <p>{{ body }}</p>
+    <article>
+      <header class="lead"><h2>{{ title }}</h2>{{ body }}</header>
+    </article>
   </div>
 </template>